fec: encoder pad the eth pkts and fec pkts
a new module pad incoming pkts up to multiple of 8. This
is needed by the FEC module since the DM creates pkts multiple
of 16bits (wr fabric width) and the FEC divides the original
frame into 4 packets. Besides, the FEC hast to pad FEC frames
under 64 bytes. Without padding an eth frame will encoded in 32 bytes
pkts which is not allow, this is why the padding up to 64 bytes