fec: add full muxtiplex for enc/dec streams
the WR Fabric sink or source is mux for letting
passing through the FEC encoded and non encoded
frames. The classification of the stream of frames
is done by a Fabric Mux which uses the field class
of the c_WRF_STATUS. Actually it multiplexes FEC
stream and EB non encoded stream. It is possible
adding new streams in combination of encoder or
decoders